[ISSUE-884] device or resource busy when NodeStageVolume

+++ pkg/base/k8s/fake_kubeclient.go
@@ -38,11 +38,23 @@ type contextKey string
 const (
 	updateFailCtxKey   = contextKey("updateFail")
 	updateFailCtxValue = "yes"
+	listFailCtxKey     = contextKey("listFail")
+	listFailCtxValue   = "yes"
+	getFailCtxKey      = contextKey("getFail")
+	getFailCtxValue    = "yes"
+	deleteFailCtxKey   = contextKey("deleteFail")
+	deleteFailCtxValue = "yes"
 )
 
 var (
 	// UpdateFailCtx raises an error on fakeClient.Update
 	UpdateFailCtx = context.WithValue(context.Background(), updateFailCtxKey, updateFailCtxValue)
+	// ListFailCtx raises an error on fakeClient.List
+	ListFailCtx = context.WithValue(context.Background(), listFailCtxKey, listFailCtxValue)
+	// GetFailCtx raises an error on fakeClient.Get
+	GetFailCtx = context.WithValue(context.Background(), getFailCtxKey, getFailCtxValue)
+	// DeleteFailCtx raises an error on fakeClient.Get
+	DeleteFailCtx = context.WithValue(context.Background(), deleteFailCtxKey, deleteFailCtxValue)
 )
 
 // NewFakeClientWrapper return new instance of FakeClientWrapper
@@ -62,6 +74,9 @@ type FakeClientWrapper struct {
 
 // Get is a wrapper around Get method
 func (fkw *FakeClientWrapper) Get(ctx context.Context, key k8sCl.ObjectKey, obj k8sCl.Object) error {
+	if ctx.Value(getFailCtxKey) == getFailCtxValue {
+		return errors.New("raise error in get")
+	}
 	if fkw.shouldPatchNS(obj) {
 		key = fkw.removeNSFromObjKey(key)
 	}
@@ -70,6 +85,9 @@ func (fkw *FakeClientWrapper) Get(ctx context.Context, key k8sCl.ObjectKey, obj 
 
 // List is a wrapper around List method
 func (fkw *FakeClientWrapper) List(ctx context.Context, list k8sCl.ObjectList, opts ...k8sCl.ListOption) error {
+	if ctx.Value(listFailCtxKey) == listFailCtxValue {
+		return errors.New("raise list error")
+	}
 	if fkw.shouldPatchNS(list) {
 		opts = fkw.removeNSFromListOptions(opts)
 	}
@@ -83,6 +101,9 @@ func (fkw *FakeClientWrapper) Create(ctx context.Context, obj k8sCl.Object, opts
 
 // Delete is a wrapper around Delete method
 func (fkw *FakeClientWrapper) Delete(ctx context.Context, obj k8sCl.Object, opts ...k8sCl.DeleteOption) error {
+	if ctx.Value(deleteFailCtxKey) == updateFailCtxValue {
+		return errors.New("raise delete error")
+	}
 	return fkw.client.Delete(ctx, obj, opts...)
 }
